+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Fix remaining time not saving to the markdown file
 * Numerous other fixes
 * Ensure no tasks are scheduled prior to 10:00 or during any meetings
+* Add the amount of tasks to the daily log output
 
 # 1.1.6: 2024-10-20